requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
aiofiles==23.2.1